messing with the code as a group to make it work properly.
