+++ app/src/main/AndroidManifest.xml
@@ -21,29 +21,12 @@
         <activity
             android:name=".HomeActivity"
             android:label="@string/title_activity_home"
-
             android:theme="@style/AppTheme.NoActionBar">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
         </activity>
         <activity android:name=".FoodActivity" />
         <activity android:name=".OpenGordonBreakfastActivity" />
-        <activity android:name=".OpenJamieBreakfast1Activity">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
+        <activity android:name=".OpenJamieBreakfast1Activity"/>
         <activity android:name=".OpenJamieBreakfast2Activity">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
         </activity>
             android:theme="@style/AppTheme.NoActionBar"/>
         <activity android:name=".WorkoutActivity"></activity>
