+++ app/src/main/java/edu/tacoma/uw/css/thuv/gethealthyapplication/MainActivity.java
@@ -19,13 +19,7 @@ public class MainActivity extends AppCompatActivity implements
 
 
 
-    @Override
-    public void addUser(String url){
-        AddUserTask task = new AddUserTask();
-        task.execute(new String[]{url.toString()});
 
-        getSupportFragmentManager().popBackStackImmediate();
-    }
 
 
     public MainActivity() {
@@ -42,8 +36,6 @@ public class MainActivity extends AppCompatActivity implements
                     .add(R.id.main_activity, new SigninFragment())
                     .commit();
         }
-
-
     }
 
     @Override
@@ -51,57 +43,9 @@ public class MainActivity extends AppCompatActivity implements
 
     }
 
-    private class AddUserTask extends AsyncTask<String, Void, String>{
 
-        @Override
-        protected void onPreExecute(){
-            super.onPreExecute();
-        }
+    @Override
+    public void addUser(String url) {
 
-        @Override
-        protected String doInBackground(String... urls){
-            String response = "";
-            HttpURLConnection urlConnection = null;
-            for (String url : urls){
-                try{
-                    URL urlObject = new URL(url);
-                    urlConnection = (HttpURLConnection) urlObject.openConnection();
-
-                    InputStream content = urlConnection.getInputStream();
-
-                    BufferedReader buffer = new BufferedReader(new InputStreamReader(content));
-                    String s = "";
-                    while((s = buffer.readLine()) != null){
-                        response += s;
-                    }
-                }catch(Exception e){
-                    response = "Unable to register. Reason: "
-                            + e.getMessage();
-                }finally {
-                    if (urlConnection != null){
-                        urlConnection.disconnect();
-                    }
-                }
-            }
-            return response;
-        }
-
-        @Override
-        protected void onPostExecute(String result){
-            try{
-                JSONObject jsonObject = new JSONObject(result);
-                String status = (String) jsonObject.get("result");
-                if(status.equals("success")){
-                    Toast.makeText(getApplicationContext(), "Successfully Registered!",
-                            Toast.LENGTH_LONG).show();
-                }else{
-                    Toast.makeText(getApplicationContext(), "Failed to register",
-                            Toast.LENGTH_LONG).show();
-                }
-            }catch(JSONException e){
-                Toast.makeText(getApplicationContext(), "Something wrong with the data" +
-                        e.getMessage(), Toast.LENGTH_LONG).show();
-            }
-        }
     }
 }
