+++ app/src/main/java/edu/tacoma/uw/css/thuv/gethealthyapplication/SigninFragment.java
@@ -3,6 +3,7 @@ package edu.tacoma.uw.css.thuv.gethealthyapplication;
 import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
+import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
@@ -14,29 +15,30 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
 
+import org.json.JSONException;
+import org.json.JSONObject;
+
+import java.io.BufferedReader;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.net.HttpURLConnection;
+import java.net.URL;
+
 
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
  * {@link SigninFragment.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link SigninFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
 public class SigninFragment extends Fragment{
-    // TODO: Rename parameter arguments, choose names that match
-    // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-    private static final String ARG_PARAM1 = "param1";
-    private static final String ARG_PARAM2 = "param2";
 
 
     private EditText mEmail;
     private EditText mPassword;
     private Button mSignIn;
-
-    // TODO: Rename and change types of parameters
-    private String mParam1;
-    private String mParam2;
+    private Button mCreateNewAccount;
 
     private OnFragmentInteractionListener mListener;
 
@@ -44,68 +46,21 @@ public class SigninFragment extends Fragment{
         // Required empty public constructor
     }
 
-    /**
-     * Use this factory method to create a new instance of
-     * this fragment using the provided parameters.
-     *
-     * @param param1 Parameter 1.
-     * @param param2 Parameter 2.
-     * @return A new instance of fragment SigninFragment.
-     */
-    // TODO: Rename and change types and number of parameters
-    public static SigninFragment newInstance(String param1, String param2) {
-        SigninFragment fragment = new SigninFragment();
-        Bundle args = new Bundle();
-        args.putString(ARG_PARAM1, param1);
-        args.putString(ARG_PARAM2, param2);
-        fragment.setArguments(args);
-        return fragment;
-    }
-
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        if (getArguments() != null) {
-            mParam1 = getArguments().getString(ARG_PARAM1);
-            mParam2 = getArguments().getString(ARG_PARAM2);
-        }
 
-        mEmail = (EditText) getActivity().findViewById(R.id.email);
-        mPassword = (EditText) getActivity().findViewById(R.id.password);
-        mSignIn = (Button) getActivity().findViewById(R.id.sign_in);
-
-        mSignIn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                String email = mEmail.getText().toString();
-                String password = mPassword.getText().toString();
-
-                if (TextUtils.isEmpty(email) || !email.contains("@")) {
-                    Toast.makeText(v.getContext(), "Enter valid email address",
-                            Toast.LENGTH_SHORT).show();
-
-                    mEmail.requestFocus();
-                } else if (TextUtils.isEmpty(password) ||
-                        password.length() < 6) {
-
-                    Toast.makeText(v.getContext(), "Enter valid password"
-                                    + " (at leaste 6 characters)",
-                            Toast.LENGTH_SHORT).show();
-
-                    mEmail.requestFocus();
-                } else {
-                    login(email, password);
-                }
-            }
-        });
     }
 
 
 
     public void login(final String theEmail, final String thePassword) {
+
+
+
         Intent i = new Intent(getActivity(), HomeActivity.class);
         startActivity(i);
-        getActivity().finish();
+        getActivity();
     }
 
     @Override
@@ -114,8 +69,37 @@ public class SigninFragment extends Fragment{
         // Inflate the layout for this fragment
         View v = inflater.inflate(R.layout.fragment_signin, container, false);
 
-        Button btn_create = (Button) v.findViewById(R.id.create_new_account);
-        btn_create.setOnClickListener(new View.OnClickListener() {
+        mEmail = (EditText) getActivity().findViewById(R.id.email);
+        mPassword = (EditText) getActivity().findViewById(R.id.password);
+        mSignIn = (Button) getActivity().findViewById(R.id.sign_in);
+
+//        mSignIn.setOnClickListener(new View.OnClickListener() {
+//            @Override
+//            public void onClick(View v) {
+//                String email = mEmail.getText().toString();
+//                String password = mPassword.getText().toString();
+//
+//                if (TextUtils.isEmpty(email) || !email.contains("@")) {
+//                    Toast.makeText(v.getContext(), "Enter valid email address",
+//                            Toast.LENGTH_SHORT).show();
+//
+//                    mEmail.requestFocus();
+//                } else if (TextUtils.isEmpty(password) ||
+//                        password.length() < 6) {
+//
+//                    Toast.makeText(v.getContext(), "Enter valid password"
+//                                    + " (at leaste 6 characters)",
+//                            Toast.LENGTH_SHORT).show();
+//
+//                    mEmail.requestFocus();
+//                } else {
+//                    login(email, password);
+//                }
+//            }
+//        });
+
+        mCreateNewAccount = (Button) v.findViewById(R.id.create_new_account);
+        mCreateNewAccount.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
                 launchRegister(v);
@@ -149,6 +133,72 @@ public class SigninFragment extends Fragment{
         }
     }
 
+    public void addUser(String url){
+        AddUserTask task = new AddUserTask();
+        task.execute(new String[]{url.toString()});
+
+        getActivity().getSupportFragmentManager().popBackStackImmediate();
+    }
+
+    private class AddUserTask extends AsyncTask<String, Void, String> {
+
+        @Override
+        protected void onPreExecute(){
+            super.onPreExecute();
+        }
+
+        @Override
+        protected String doInBackground(String... urls){
+            String response = "";
+            HttpURLConnection urlConnection = null;
+            for (String url : urls){
+                try{
+                    URL urlObject = new URL(url);
+                    urlConnection = (HttpURLConnection) urlObject.openConnection();
+
+                    InputStream content = urlConnection.getInputStream();
+
+                    BufferedReader buffer = new BufferedReader(new InputStreamReader(content));
+                    String s = "";
+                    while((s = buffer.readLine()) != null){
+                        response += s;
+                    }
+                }catch(Exception e){
+                    response = "Unable to register. Reason: "
+                            + e.getMessage();
+                }finally {
+                    if (urlConnection != null){
+                        urlConnection.disconnect();
+                    }
+                }
+            }
+            return response;
+        }
+
+        @Override
+        protected void onPostExecute(String result){
+            try{
+                JSONObject jsonObject = new JSONObject(result);
+                String status = (String) jsonObject.get("result");
+                if(status.equals("success")){
+                    Toast.makeText(getActivity().getApplicationContext(), "Successfully Registered!",
+                            Toast.LENGTH_LONG).show();
+                }else{
+                    Toast.makeText(getActivity().getApplicationContext(), "Failed to register",
+                            Toast.LENGTH_LONG).show();
+                }
+            }catch(JSONException e){
+                Toast.makeText(getActivity().getApplicationContext(), "Something wrong with the data" +
+                        e.getMessage(), Toast.LENGTH_LONG).show();
+            }
+        }
+    }
+
+
+
+
+
+
     @Override
     public void onDetach() {
         super.onDetach();
@@ -169,4 +219,5 @@ public class SigninFragment extends Fragment{
         // TODO: Update argument type and name
         void onFragmentInteraction(Uri uri);
     }
+
 }
