+++ app/src/main/res/layout/fragment_registration.xml
@@ -53,7 +53,7 @@
 
     <EditText
         android:id="@+id/text_user_email"
-        android:layout_width="wrap_content"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:hint="@string/email"
         android:inputType="textEmailAddress" />
